+++ package.json
@@ -4,7 +4,7 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
+    "start": "node server.js", 
     "seed": "node seeds/seed.js",
     "watch": "nodemon server.js"
   },
