+++ public/src/app/app.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from "@angular/core";
+
+import { AppComponent } from "./app.components.spec";
+import { HttpClientModule } from "@angular/common/http";
+import { BrowserModule } from '@angular/platform-browser';
+
+@NgModule({
+    declarations: [
+        AppComponent
+    ],
+    imports: [
+        BrowserModule,
+        HttpClientModule
+    ],
+    providers: [],
+    bootstrap: [AppComponent]
+})
+
+export class AppModule { }
